move command tests to separate modules
